% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latla_funcs.R
\name{latla_thres}
\alias{latla_thres}
\title{Thresholding function of LATLA}
\usage{
latla_thres(pvs, pis, ws, q)
}
\arguments{
\item{pvs}{Vector of p-values under the null hypotheses.}

\item{pis}{The sparsity levels.}

\item{ws}{Weights for the p-values.}

\item{q}{FDR level.}
}
\value{
A list of three: 1.number of rejections; 2.rejection threshold;
3.indication of whether each index is rejected.
}
\description{
This function first ranks all hypotheses according to the weighted
p-values and then determines a threshold along the ranking to control
the FDR. See algorithm 1 in reference paper for more details.
}
