% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latla_funcs.R
\name{latla_weights}
\alias{latla_weights}
\title{Data-driven weights of LATLA}
\usage{
latla_weights(x, d, pis, mu0, sd0, q = 0.05, h = "auto", eps = 0.1)
}
\arguments{
\item{x}{Vector of the primary statistics.}

\item{d}{Distance matrix.}

\item{pis}{Estimated sparsity level.}

\item{mu0}{Vector of means under the null hypotheses.}

\item{sd0}{Vector of standard deviations under the null hypotheses.}

\item{q}{FDR level, default value is set to be 0.05.}

\item{h}{Bandwidth for kernel estimation, default is set to 'auto', where
the bandwidth is automatically selected by build-in method 'ste'.}

\item{eps}{Determines the size of the neighborhood, only m^(1-eps) data
points will be used for kernel estimation. eps should range
from 0 to 1.}
}
\value{
The oracle-assisted weights.
}
\description{
The function calculates the oracle-assisted data-driven weights described
in the algorithm 2 in LATLA paper.
}
